rename  DefaultLogRecordBatch to MemoryLogRecordBatch
