+++ fluss-common/src/main/java/com/alibaba/fluss/record/DefaultLogRecord.java
@@ -29,7 +29,7 @@ import com.alibaba.fluss.utils.MurmurHashUtils;
 
 import java.io.IOException;
 
-import static com.alibaba.fluss.record.DefaultLogRecordBatch.LENGTH_LENGTH;
+import static com.alibaba.fluss.record.MemoryLogRecordBatch.LENGTH_LENGTH;
 
 /* This file is based on source code of Apache Kafka Project (https://kafka.apache.org/), licensed by the Apache
  * Software Foundation (ASF) under the Apache License, Version 2.0. See the NOTICE file distributed with this work for
